Refactor SideBar
